refactor(View): split out Parented and Single View
